+++ tcod/bsp.py
@@ -148,20 +148,17 @@ class BSP:
         """Divide this partition recursively.
 
         Args:
-            depth (int): The maximum depth to divide this object recursively.
-            min_width (int): The minimum width of any individual partition.
-            min_height (int): The minimum height of any individual partition.
-            max_horizontal_ratio (float):
-                Prevent creating a horizontal ratio more extreme than this.
-            max_vertical_ratio (float):
-                Prevent creating a vertical ratio more extreme than this.
-            seed (Optional[tcod.random.Random]):
-                The random number generator to use.
+            depth: The maximum depth to divide this object recursively.
+            min_width: The minimum width of any individual partition.
+            min_height: The minimum height of any individual partition.
+            max_horizontal_ratio: Prevent creating a horizontal ratio more extreme than this.
+            max_vertical_ratio: Prevent creating a vertical ratio more extreme than this.
+            seed: The random number generator to use.
         """
         cdata = self._as_cdata()
         lib.TCOD_bsp_split_recursive(
             cdata,
-            seed or ffi.NULL,
+            seed.random_c if seed is not None else ffi.NULL,
             depth,
             min_width,
             min_height,